Completed implementation of PID controller in PID.cpp. Removed redundant files.
