--- include/lib.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include<iostream>
-
-void dummy()
-{
-    std::cout << "HI" << std::endl;
-}
